Update data about Web Annotation joint deliverable
This PR updates the information about the joint deliverable (Robust Anchoring) with the Web Annotation WG and adds that WG to list of liaisons. 

(WebAnnotWG has not yet published a TR for this deliverable, although the have started work on "RangeFinder API"; see <https://www.w3.org/annotation/wiki/PubStatus>).