Microdata doesn't have a DOM API
fix #157

There is no obvious justification for a microdata DOM API, so whether it is in scope or not is irrelevant. On the other hand if browsers suddenly implemented one, e.g. based on the old specification for it, then it would be silly not to have it in scope.